去掉是否是用户资源； 前端合二为一； 修复若干bug；

+++ randl_service/src/main/java/com/jeramtough/randl2/service/oauth/impl/OauthResourceDetailsServiceImpl.java
@@ -145,7 +145,9 @@ public class OauthResourceDetailsServiceImpl extends MyBaseServiceImpl<OauthReso
                 .parallelStream()
                 .map(OauthScopeDetailsDto::getFid)
                 .collect(Collectors.toList());
-        oauthScopeDetailsService.removeByIds(removeIds);
+        if (!removeIds.isEmpty()){
+            oauthScopeDetailsService.removeByIds(removeIds);
+        }
 
         return "更新成功！";
     }
